chore(canvas) Moved ModeSelectButtons into a separate file
